company service 수정 (#45)
* 시간별 시세 크롤링

* Feature 시세

* 1

* 1

* solve conflict

* feature/POST_CRUD

* 마지막 업데이트 날짜부터 데이터 업데이트

* feature/Company_info 기능 수정

+++ src/main/java/com/KWdatabase/teamProject/Model/Company.java
@@ -13,7 +13,7 @@ import java.time.LocalDateTime;
 public class Company {
     private String itemCode;
     private String companyName;
-    private int itemNumber;
+    private long itemNumber;
     private String companySummary;
     private LocalDate publicDate;
 }
